+++ APIProjecteKanban/DAL/Model/Task.cs
@@ -2,14 +2,14 @@
 {
     public class Task
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty;
         public string Priority { get; set; } = string.Empty;
-        public int IdResponsible { get; set; }
+        public long IdResponsible { get; set; }
         public DateTime CreationDate { get; set; }
         public DateTime? LastUpdate { get; set; }
-        public int IdProject { get; set; }
-        public int State { get; set; }
+        public long IdProject { get; set; }
+        public long State { get; set; }
     }
 }
